Refactor server structure and add weather chatbot route
Reorganized server code by removing root-level .gitignore, package.json, and lock files, and moving dependencies and configuration into the server directory. Added a dedicated chatBot.js route for weather queries, with weather logic extracted to utils/getWeather.js. Cleaned up Auth.js to focus on authentication, and updated app.js to use the new routes. Updated dependencies to include axios and node-fetch for API calls.
