--- server/.env
@@ -1,6 +1,0 @@
-PORT=5000
-MONGO_URI=mongodb+srv://suriyan:COOWG9nsnloxnMtV@cluster0.pzuq2ov.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0
-WEATHER_API_KEY=1a4fd26a464005c108d133b93d44e23f
-OPENAI_API_KEY=sk-2947d57f93e548dc98226844681a171d
-REACT_APP_OPENAI_API_KEY=sk-2947d57f93e548dc98226844681a171d
-REACT_APP_RAPIDAPI_KEY=02e74478damsha804b27d0a44a83p14a316jsn1545ead794ed
