+++ server/routes/Auth.js
@@ -1,89 +1,81 @@
+const User = require('../models/User');
 const express = require('express');
-const axios = require('axios');
 const router = express.Router();
 
-const OPENWEATHER_API_KEY = process.env.OPENWEATHER_API_KEY; // Store in .env
-
-// Helper: Get current weather
-async function getCurrentWeather(city) {
-  const url = `https://api.openweathermap.org/data/2.5/weather?q=${city}&appid=${OPENWEATHER_API_KEY}&units=metric`;
-  const response = await axios.get(url);
-  return `The current temperature in ${city} is ${response.data.main.temp}°C with ${response.data.weather[0].description}.`;
-}
-
-// Helper: Get tomorrow's weather
-async function getTomorrowWeather(city) {
-  const url = `https://api.openweathermap.org/data/2.5/forecast?q=${city}&appid=${OPENWEATHER_API_KEY}&units=metric`;
-  const response = await axios.get(url);
-  const tomorrow = new Date();
-  tomorrow.setDate(tomorrow.getDate() + 1);
-  const tomorrowDate = tomorrow.toISOString().split('T')[0];
-
-  const tomorrowData = response.data.list.filter(item =>
-    item.dt_txt.startsWith(tomorrowDate)
-  );
-
-  if (tomorrowData.length === 0) {
-    return `No forecast data available for tomorrow in ${city}.`;
-  }
-
-  const avgTemp =
-    tomorrowData.reduce((sum, item) => sum + item.main.temp, 0) /
-    tomorrowData.length;
-
-  return `Tomorrow's average temperature in ${city} will be around ${avgTemp.toFixed(1)}°C with ${tomorrowData[0].weather[0].description}.`;
-}
-
-// Helper: Multi-day forecast
-async function getMultiDayForecast(city, days = 3) {
-  const url = `https://api.openweathermap.org/data/2.5/forecast?q=${city}&appid=${OPENWEATHER_API_KEY}&units=metric`;
-  const response = await axios.get(url);
-
-  const forecasts = {};
-  response.data.list.forEach(item => {
-    const date = item.dt_txt.split(' ')[0];
-    if (!forecasts[date]) forecasts[date] = [];
-    forecasts[date].push(item.main.temp);
-  });
-
-  const today = new Date().toISOString().split('T')[0];
-  const dates = Object.keys(forecasts).filter(d => d !== today).slice(0, days);
-
-  let forecastMsg = `Weather forecast for the next ${days} days in ${city}:\n`;
-  dates.forEach(date => {
-    const avg =
-      forecasts[date].reduce((sum, t) => sum + t, 0) / forecasts[date].length;
-    forecastMsg += `${date}: ~${avg.toFixed(1)}°C\n`;
-  });
-
-  return forecastMsg.trim();
-}
-
-// Main chatbot route
-router.post('/chat', async (req, res) => {
+router.post('/signup', async (req, res) => {
+  const { username, password } = req.body;
   try {
-    const userMessage = req.body.message.toLowerCase();
-    let reply;
+    const existingUser = await User.findOne({ username });
+    if (existingUser) return res.status(400).json({ message: "Username already taken" });
 
-    // Parse city and intent
-    const cityMatch = userMessage.match(/in ([a-zA-Z\s]+)/);
-    const city = cityMatch ? cityMatch[1].trim() : null;
+    const newUser = new User({ username, password });
+    await newUser.save();
 
-    if (!city) {
-      reply = "Please tell me a city, e.g., 'weather in Chennai'.";
-    } else if (userMessage.includes('tomorrow')) {
-      reply = await getTomorrowWeather(city);
-    } else if (userMessage.includes('next') || userMessage.includes('few days')) {
-      reply = await getMultiDayForecast(city, 3);
-    } else {
-      reply = await getCurrentWeather(city);
-    }
-
-    res.json({ reply });
+    res.status(201).json({ message: "Signup successful" });
   } catch (error) {
-    console.error(error);
-    res.json({ reply: "Sorry, I couldn't fetch the weather right now." });
+    res.status(500).json({ error: "Signup failed" });
   }
 });
 
+router.post('/login', async (req, res) => {
+  const { username, password } = req.body;
+  try {
+    const user = await User.findOne({ username, password });
+    if (!user) return res.status(400).json({ message: "Invalid credentials" });
+
+    res.status(200).json({ username: user.username, message: "Login successful" });
+  } catch (error) {
+    res.status(500).json({ error: "Login failed" });
+  }
+});
+
+
 module.exports = router;
+
+
+// // Check if username exists
+// const checkUsername = async (req, res) => {
+//   const { username } = req.params;
+
+//   try {
+//     const user = await User.findOne({ username });
+//     if (user) {
+//       return res.status(200).json({ available: false, message: "Username is taken" });
+//     } else {
+//       return res.status(200).json({ available: true, message: "Username is available" });
+//     }
+//   } catch (error) {
+//     return res.status(500).json({ available: false, message: "Server error" });
+//   }
+// };
+
+// const signup = async (req, res) => {
+//   const { username, password } = req.body;
+
+//   try {
+//     const existingUser = await User.findOne({ username });
+//     if (existingUser) return res.status(400).json({ message: "Username already taken" });
+
+//     const newUser = new User({ username, password });
+//     await newUser.save();
+
+//     res.status(201).json({ message: "Signup successful" });
+//   } catch (error) {
+//     res.status(500).json({ error: "Signup failed" });
+//   }
+// };
+
+// const login = async (req, res) => {
+//   const { username, password } = req.body;
+
+//   try {
+//     const user = await User.findOne({ username, password });
+//     if (!user) return res.status(400).json({ message: "Invalid credentials" });
+
+//     res.status(200).json({ username: user.username });
+//   } catch (error) {
+//     res.status(500).json({ error: "Login failed" });
+//   }
+// };
+
+// module.exports = { signup, login, checkUsername };
